+++ infrastructure/geom/CoordinateAxis/new_CoordinateAxis.F90
@@ -7,13 +7,15 @@ submodule (mapl_CoordinateAxis_mod) new_CoordinateAxis_smod
 
 contains
    
-   pure module function new_CoordinateAxis(centers, corners) result(axis)
+   pure module function new_CoordinateAxis(centers, corners, tolerance) result(axis)
       type(CoordinateAxis) :: axis
       real(kind=R8), intent(in) :: centers(:)
       real(kind=R8), intent(in) :: corners(:)
+      real(kind=R8), optional, intent(in) :: tolerance
 
       axis%centers = centers
       axis%corners = corners
+      if (present(tolerance)) axis%tolerance = tolerance
    end function new_CoordinateAxis